Make basket service aware of basket draft status
Baskets which haven't been ordered should be in the draft status.
This is now happening with the basket service.  The service creates
the baskets in a draft status and only returns draft baskets when the
current basket is requested (a completed basket should no longer be
edited by the user).

We suspect other services may move these non-draft baskets around in
order to ensure that they cannot be changed anymore.  That is not the
goal of the current service se we ignore that situation for now.  The
final implementation may change to cope with that.