Correct basket registration for logged in users
Users which were logged in were not correctly detected in the checkout
phase of the basket.  The views report a positive result but the basket
itself was not marked as delivered due to incorrect selection of the
actor to whom the basket should belong.

Because each actor has their own graph with their baskets in it, we can
ignore the selection of the actor itself, and select any baskets that
match instead.