Use graph to filter out basket rather than session id
The session id only works for users which have not logged in.  The graph
as a source can be used both when the user is logged in (in which case
it is the user uri) as well as when the user chooses not to make an
account (in which case it is the session id).